+++ src/plugins/drag-and-drop/drag-and-drop.ts
@@ -20,8 +20,8 @@ import { Dom } from 'jodit/core/dom/dom';
 import { attr, ctrlKey, dataBind, getDataTransfer } from 'jodit/core/helpers';
 import { Plugin } from 'jodit/core/plugin';
 import { autobind, throttle } from 'jodit/core/decorators';
-import { isFileBrowserFilesItem } from 'jodit/modules/file-browser';
 import { pluginSystem } from 'jodit/core/global';
+import { FileBrowserFiles } from "jodit/modules/file-browser/ui";
 
 /**
  * Process drag and drop image from FileBrowser and movev image inside the editor
@@ -226,6 +226,18 @@ export class dragAndDrop extends Plugin {
 				this.onDragStart
 			);
 	}
+}
+
+/**
+ * @private
+ */
+function isFileBrowserFilesItem(target: unknown): boolean {
+	return (
+		Dom.isElement(target) &&
+		target.classList.contains(
+			FileBrowserFiles.prototype.getFullElName('item')
+		)
+	);
 }
 
 pluginSystem.add('dragAndDrop', dragAndDrop);